updated notebook and script
