+++ AI-and-Analytics/Features-and-Functionality/IntelPyTorch_Extensions_Inference_Optimization/requirements.txt
@@ -1,4 +1,3 @@
 notebook
-opencv-python
 matplotlib
 transformers
